// script/types.ts
export interface MarkdownFrontMatter {
    /** Article title */
    title: string;

    /** Whether to show comments */
    comment?: boolean;

    /** Whether to show timestamp, i.e. Creation & Modification time */
    timestamp?: boolean;
}
// ...
export interface MarkdownHeaderJsx {
    /** Header level */
    level: number;

    /** Header title */
    title: JSX.Element;

    /** Header permalink */
    link: string;
}
// ...
export interface RouteMeta {
    /** Pathname of local markdown file */
    pathname: string;

    /** Category of the route, empty string for `/index` */
    category: string;

    /** Body component */
    body: () => Promise<{ default: () => JSX.Element }>;

    /** Parsed front matter */
    attr: MarkdownFrontMatter;

    /** Table of contents */
    toc: MarkdownHeaderJsx[];

    /** Creation time */
    created: string;

    /** Last modified time */
    updated: string;

    /** Page type */
    type: "index" | "post" | "404" | "root";

    /** Back link */
    back: { title: string; link: string };

    /** Route scroll to */
    scrollTo?: { left: number; top: number };
}